fixed up the odd line endings, moved to subfolder
