+++ python/example_code/iot/plantwateringsample/simulate-random-moisture-levels.py
@@ -12,7 +12,7 @@
 # specific language governing permissions and limitations under the License.
 #
 
-# snippet-sourcedescription:[read-sensor-send-to-iot.py demonstrates how to simulate soil moisture readings by generating random data. You push that data into the related shadow in AWS IoT.]
+# snippet-sourcedescription:[simulate-random-moisture-levels.py demonstrates how to simulate soil moisture readings by generating random data. You push that data into the related shadow in AWS IoT.]
 # snippet-service:[iot]
 # snippet-keyword:[Python]
 # snippet-keyword:[AWS IoT]
@@ -20,150 +20,76 @@
 # snippet-sourcetype:[full-example]
 # snippet-sourcedate:[2019-01-25]
 # snippet-sourceauthor:[pccornel]
-# snippet-start:[iot.python.read-sensor-send-to-iot.complete]
-
+# snippet-start:[iot.python.simulate_moisture_levels.complete]
 
 from AWSIoTPythonSDK.MQTTLib import AWSIoTMQTTShadowClient
-
-import RPi.GPIO as GPIO
-
-import time
-
-
+import random, time
 
 # A random programmatic shadow client ID.
-
 SHADOW_CLIENT = "myShadowClient"
 
-
-
-# The unique hostname that AWS IoT generated for 
-
+# The unique hostname that &IoT; generated for 
 # this device.
-
 HOST_NAME = "yourhostname-ats.iot.us-east-1.amazonaws.com"
 
-
-
-# The relative path to the correct root CA file for AWS IoT, 
-
-# that you have already saved onto this device.
-
+# The relative path to the correct root CA file for &IoT;, 
+# which you have already saved onto this device.
 ROOT_CA = "AmazonRootCA1.pem"
 
-
-
 # The relative path to your private key file that 
-
-# AWS IoT generated for this device, that you 
-
+# &IoT; generated for this device, which you 
 # have already saved onto this device.
-
 PRIVATE_KEY = "yourkeyid-private.pem.key"
 
-
-
 # The relative path to your certificate file that 
-
-# AWS IoT generated for this device, that you 
-
+# &IoT; generated for this device, which you 
 # have already saved onto this device.
-
 CERT_FILE = "yourkeyid-certificate.pem.crt.txt"
 
-
-
 # A programmatic shadow handler name prefix.
-
 SHADOW_HANDLER = "MyRPi"
 
-
-
 # Automatically called whenever the shadow is updated.
-
 def myShadowUpdateCallback(payload, responseStatus, token):
-
   print()
-
-  print('UPDATE: $aws/things/' + SHADOW_HANDLER +
- 
+  print('UPDATE: $aws/things/' + SHADOW_HANDLER + 
     '/shadow/update/#')
-
   print("payload = " + payload)
-
   print("responseStatus = " + responseStatus)
-
   print("token = " + token)
 
-
-
 # Create, configure, and connect a shadow client.
-
 myShadowClient = AWSIoTMQTTShadowClient(SHADOW_CLIENT)
-
 myShadowClient.configureEndpoint(HOST_NAME, 8883)
-
 myShadowClient.configureCredentials(ROOT_CA, PRIVATE_KEY,
-
   CERT_FILE)
-
 myShadowClient.configureConnectDisconnectTimeout(10)
-
 myShadowClient.configureMQTTOperationTimeout(5)
-
 myShadowClient.connect()
 
-
-
 # Create a programmatic representation of the shadow.
-
 myDeviceShadow = myShadowClient.createShadowHandlerWithName(
-
   SHADOW_HANDLER, True)
 
-
-
-# Represents the GPIO21 pin on the Raspberry Pi.
-
-channel = 21
-
-
-
-# Use the GPIO BCM pin numbering scheme.
-
-GPIO.setmode(GPIO.BCM)
-
-
-
-# Receive input signals through the pin.
-
-GPIO.setup(channel, GPIO.IN)
-
-
-
+# Keep generating random test data until this script 
+# stops running.
+# To stop running this script, press Ctrl+C.
 while True:
+  # Generate random True or False test data to represent
+  # okay or low moisture levels, respectively.
+  moisture = random.choice([True, False])
 
- 
- if GPIO.input(channel):
-
-     myDeviceShadow.shadowUpdate(
-
-       '{"state":{"reported":{"moisture":"low"}}}',
- 
+  if moisture:
+    myDeviceShadow.shadowUpdate(
+      '{"state":{"reported":{"moisture":"okay"}}}',
       myShadowUpdateCallback, 5)
-
   else:
-
-     myDeviceShadow.shadowUpdate(
-
-       '{"state":{"reported":{"moisture":"okay"}}}',
- 
+    myDeviceShadow.shadowUpdate(
+      '{"state":{"reported":{"moisture":"low"}}}',
       myShadowUpdateCallback, 5)
-
-
 
   # Wait for this test value to be added.
-
   time.sleep(60)
 
-# snippet-end:[iot.python.read-sensor-send-to-iot.complete]
+# snippet-end:[iot.python.simulate_moisture_levels.complete]
+
